refactor dependencies on ep-eth crate
