+++ frame/ethink/Cargo.toml
@@ -6,7 +6,7 @@ description = "Ethereum RPC compatibility aid for pallet-contracts. This crate i
 edition.workspace = true
 
 [dependencies]
-ethereum = { workspace = true, features = ["with-codec"] }
+ep-eth.workspace = true
 ethereum-types.workspace = true
 scale-codec.workspace = true
 scale-info.workspace = true
@@ -51,6 +51,9 @@ std = [
 	"sp-io/std",
 	"sp-runtime/std",
 	"sp-std/std",
+	# ethink
+	"ep-eth/std",
+
 ]
 runtime-benchmarks = [
 	"frame-support/runtime-benchmarks",
